Add build dir
Adding the build directory. We have since been relying on github
pages to handle deployments, but this comes at a cost of having
different environments. Case in point, github-pages cannot use
jekyll-archives, which means the way we have done categories/tags
does not work.

We can instead simple serve the same site that we build locally here.
We just have to add the actual build dir, _site (renamed to docs),
and deploy from that.